Penalty Scores
Reduce the score of the template for each word in the template
that is not matched. This is to prevent Naomi from trying to
match one very common word, like "WHAT" or "IS".

The problem with this approach is that it is difficult to convert
the score into a confidence percent, and it is easy for the score
to go negative. Since Naomi is looking for a confidence of 5%,
anything lower than .05 (which includes any negative score) is
rejected. Also, two templates that are functionally similar can
end up with different results and are not seen as alternatives.

For example, if you have the following templates:
Tell me a knock knock joke
Tell me a dad joke
Tell me a chuck norris joke

and the user utters:
Tell me a joke

Then I would want Naomi to match all three of the above and choose
one type of joke randomly. But when adding negative scores dad
joke would win every time because it has only one word missing
while the other two have two words missing.